ref: Исправлено подсвечивание ответа в режиме ввода ответа и изменено сообщение ошибки авторизации
