+++ data/data-training/src/main/kotlin/com/example/training/data/util/GeneratePartialAnswer.kt
@@ -2,8 +2,7 @@ package com.example.training.data.util
 
 /**
  * Функция для генерации частичного ответа для режима "Дополнить ответ".
- * Генерирует строку с пропусками, где определённое количество слов заменяется на "...",
- * чтобы пользователь мог дополнить их в процессе тренировки.
+ * Генерирует строку с пропусками, где определённое количество слов заменяется на "_", чтобы пользователь мог дополнить их в процессе тренировки.
  *
  * Стратегия генерации пропусков:
  * - Пропускаются случайные слова в ответе, но не более, чем определённый процент от общего количества слов,
@@ -11,15 +10,18 @@ package com.example.training.data.util
  * - Пропуски размещаются случайным образом в пределах строки, начиная с случайного индекса.
  *
  * @param answer Ответ, для которого нужно сгенерировать частичный ответ с пропусками.
- * @return Пара:
- *  - `String`: Частичный ответ, где некоторые слова заменены на "...".
+ * @return `Triple<String, List<String>, Int>`:
+ *  - `String`: Частичный ответ, где некоторые слова заменены на "_".
  *  - `List<String>`: Список пропущенных слов.
+ *  - `Int`: Индекс в ответе, с которого начинаются пропуски.
  */
-internal fun generatePartialAnswer(answer: String): Pair<String, List<String>> {
+internal fun generatePartialAnswer(answer: String): Triple<String, List<String>, Int> {
     val words = answer.split(" ")
-    if (words.size <= MIN_WORDS_FOR_BLANK) return "" to words
 
-    // Рассчитываем максимальное количество пропущенных слов
+    // Если количество слов в ответе меньше минимального для пропусков, возвращаем пустую строку и пустой список
+    if (words.size <= MIN_WORDS_FOR_BLANK) return Triple("", words, -1)
+
+    // Рассчитываем максимальное количество пропущенных слов (ограничение по проценту)
     val maxMissingWordsCount = (words.size * MAX_BLANK_PERCENT / 100).coerceIn(1, MAX_MISSING_WORDS)
 
     // Случайное количество пропущенных слов
@@ -28,18 +30,18 @@ internal fun generatePartialAnswer(answer: String): Pair<String, List<String>> {
     // Выбираем случайный индекс для начала пропусков
     val startIndex = (0..(words.size - missingWordCount)).random()
 
-    // Получаем слова, которые будут пропущены
+    // Получаем список слов, которые будут пропущены
     val missingWords = words.subList(startIndex, startIndex + missingWordCount)
 
-    // Генерация частичного ответа с пропусками
+    // Генерация частичного ответа с пропусками (заменяем слова на "_")
     val partialAnswer = words.toMutableList().apply {
         for (i in startIndex until (startIndex + missingWordCount)) {
-            //this[i] = "..."
-            this[i] = "_".repeat(this[i].length)
+            this[i] = "_".repeat(this[i].length) // Заменяем слово на пропуск
         }
     }.joinToString(" ")
 
-    return partialAnswer to missingWords
+    // Возвращаем частичный ответ, список пропущенных слов и индекс начала пропусков
+    return Triple(partialAnswer, missingWords, startIndex)
 }
 
 private const val MIN_WORDS_FOR_BLANK = 3
